simplified the binary operations stuff and replaced it with a direct binary modification by the transformer
