+++ src/Processor/AbstractProcessor.php
@@ -5,7 +5,6 @@ namespace JoliCode\MediaBundle\Processor;
 use JoliCode\MediaBundle\Binary\Binary;
 use JoliCode\MediaBundle\Model\Format;
 use JoliCode\MediaBundle\Transformation\Transformation;
-use JoliCode\MediaBundle\Transformer\BinaryOperation\BinaryOperationInterface;
 
 abstract readonly class AbstractProcessor extends AbstractProcessCreator implements ProcessorInterface
 {
@@ -19,14 +18,6 @@ abstract readonly class AbstractProcessor extends AbstractProcessCreator impleme
     public function canProcessOutputFormat(string $outputFormat): bool
     {
         return \in_array(Format::fromName($outputFormat), $this->getProcessableOutputFormats(), true);
-    }
-
-    /**
-     * @param array<string, mixed> $processingOptions
-     */
-    public function processBinaryOperation(Binary $binary, BinaryOperationInterface $binaryOperation, array $processingOptions = []): Binary
-    {
-        throw new \LogicException('The processor does not support processing binary operations.');
     }
 
     abstract public function getName(): string;
