+++ src/Processor/ProcessorInterface.php
@@ -5,7 +5,6 @@ namespace JoliCode\MediaBundle\Processor;
 use JoliCode\MediaBundle\Binary\Binary;
 use JoliCode\MediaBundle\Model\Format;
 use JoliCode\MediaBundle\Transformation\Transformation;
-use JoliCode\MediaBundle\Transformer\BinaryOperation\BinaryOperationInterface;
 
 interface ProcessorInterface
 {
@@ -29,9 +28,4 @@ interface ProcessorInterface
      * @param array<string, mixed> $processingOptions
      */
     public function process(Binary $binary, Transformation $transformation, array $processingOptions = [], ?string $forceOutputFormat = null): Binary;
-
-    /**
-     * @param array<string, mixed> $processingOptions
-     */
-    public function processBinaryOperation(Binary $binary, BinaryOperationInterface $binaryOperation, array $processingOptions = []): Binary;
 }
