+++ src/Transformer/Heighten.php
@@ -4,7 +4,7 @@ namespace JoliCode\MediaBundle\Transformer;
 
 use JoliCode\MediaBundle\Transformation\Transformation;
 
-readonly class Heighten extends AbstractTransformer implements TransformerInterface, WithTransformTransformerInterface
+readonly class Heighten extends AbstractTransformer implements TransformerInterface
 {
     /**
      * @param int|string $height
@@ -30,6 +30,5 @@ readonly class Heighten extends AbstractTransformer implements TransformerInterf
         $ratio = $height / $transformation->targetHeight;
         $transformation->targetWidth = (int) round($transformation->targetWidth * $ratio);
         $transformation->targetHeight = $height;
-        $transformation->mustRun = true;
     }
 }
