+++ src/Transformer/NeedsImmediateProcessingTransformerInterface.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace JoliCode\MediaBundle\Transformer;
+
+interface NeedsImmediateProcessingTransformerInterface
+{
+}
