+++ src/Transformer/Thumbnail.php
@@ -4,7 +4,7 @@ namespace JoliCode\MediaBundle\Transformer;
 
 use JoliCode\MediaBundle\Transformation\Transformation;
 
-readonly class Thumbnail extends AbstractTransformer implements TransformerInterface, WithTransformTransformerInterface
+readonly class Thumbnail extends AbstractTransformer implements TransformerInterface
 {
     public function __construct(
         private int $width,
@@ -58,7 +58,6 @@ readonly class Thumbnail extends AbstractTransformer implements TransformerInter
         $transformation->cropY = $currentCropY + $cropY;
         $transformation->cropWidth = $cropWidth;
         $transformation->cropHeight = $cropHeight;
-        $transformation->mustRun = true;
     }
 
     private function getCropPosition(): float
