+++ src/Transformer/TransformerInterface.php
@@ -2,6 +2,9 @@
 
 namespace JoliCode\MediaBundle\Transformer;
 
+use JoliCode\MediaBundle\Transformation\Transformation;
+
 interface TransformerInterface
 {
+    public function transform(Transformation $transformation): void;
 }
