--- src/Transformer/WithOperationTransformerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace JoliCode\MediaBundle\Transformer;
-
-use JoliCode\MediaBundle\Transformer\BinaryOperation\BinaryOperationInterface;
-
-interface WithOperationTransformerInterface
-{
-    public function getBinaryOperation(?int $binaryWidth, ?int $binaryHeight): BinaryOperationInterface;
-}
