--- src/Transformer/WithTransformTransformerInterface.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace JoliCode\MediaBundle\Transformer;
-
-use JoliCode\MediaBundle\Transformation\Transformation;
-
-interface WithTransformTransformerInterface
-{
-    public function transform(Transformation $transformation): void;
-}
